Codec.Crypto.AES.Random

Description

This module provides a cryptographically secure PRNG based on AES, reading the seed from devurandom

Documentation

randBytes :: Int -> IO ByteString

Randomness from a system source of nonsense such as devurandom

prandBytes :: Int -> IO ByteString

Cryptographic pseudorandomness from an AES cipher. This function is currently inefficient for non-multiple-of-16 sized bytestrings.

data AESGen

A random number generator that gets its input from prandBytes, ensuring purity by storing those bytes for later if you don't discard the generator.

Using split on this generator isn't supported, but could be.

Please note that if an asynchronous exception is caught while a random number is being generated, the generator will be wrecked forevermore.

newAESGen :: IO AESGen